Update AADSyncSettings.ps1
Had some trouble with this tool for the couple last days at work. I had local admin on the AD Connect server, but the script always failed saying: "Login failed for user '' " and kept erroring out. To get it to work, I first had to:
1. Change my security context from local admin to SYSTEM.
2. Manually add "Trusted_Connection=True" to the connectionString inside the Get-AADConfigDbConnection, so that it would authenticate using the current security context, which should be SYSTEM.

Without it, it just wouldn't work.

Cheers.